Moved Suspense-specific types into src/suspense package
